Implemented basic error checking for loads and nodes.
Start node now is always the leftmost node of a beam, regardless of the order in which it was assigned to a beam.

Dimensions are now checked on all loads and other basic logic checks for point forces and distributed forces are now included.

Added code for finding the LeftMostNode and the RightMostNode of our system.